+++ android/plugin/src/main/java/com/marianhello/bgloc/Config.java
@@ -48,7 +48,10 @@ public class Config implements Parcelable
     private Boolean startForeground = true;
     private Boolean stopOnStillActivity = true;
     private String url;
+    private String syncUrl;
+    private Integer syncThreshold = 100;
     private HashMap httpHeaders = new HashMap<String, String>();
+    private Integer maxLocations = 10000;
 
     public Config () {
     }
@@ -77,6 +80,9 @@ public class Config implements Parcelable
         out.writeInt(getActivitiesInterval());
         out.writeValue(getStopOnStillActivity());
         out.writeString(getUrl());
+        out.writeString(getSyncUrl());
+        out.writeInt(getSyncThreshold());
+        out.writeInt(getMaxLocations());
         Bundle bundle = new Bundle();
         bundle.putSerializable("httpHeaders", getHttpHeaders());
         out.writeBundle(bundle);
@@ -112,6 +118,9 @@ public class Config implements Parcelable
         setActivitiesInterval(in.readInt());
         setStopOnStillActivity((Boolean) in.readValue(null));
         setUrl(in.readString());
+        setSyncUrl(in.readString());
+        setSyncThreshold(in.readInt());
+        setMaxLocations(in.readInt());
         Bundle bundle = in.readBundle();
         setHttpHeaders((HashMap<String, String>) bundle.getSerializable("httpHeaders"));
     }
@@ -262,6 +271,30 @@ public class Config implements Parcelable
         this.url = url;
     }
 
+    public Boolean hasUrl() {
+        return this.url != null && !this.url.isEmpty();
+    }
+
+    public String getSyncUrl() {
+        return syncUrl;
+    }
+
+    public void setSyncUrl(String syncUrl) {
+        this.syncUrl = syncUrl;
+    }
+
+    public Boolean hasSyncUrl() {
+        return this.syncUrl != null && !this.syncUrl.isEmpty();
+    }
+
+    public Integer getSyncThreshold() {
+        return syncThreshold;
+    }
+
+    public void setSyncThreshold(Integer syncThreshold) {
+        this.syncThreshold = syncThreshold;
+    }
+
     public HashMap getHttpHeaders() {
         return this.httpHeaders;
     }
@@ -271,6 +304,8 @@ public class Config implements Parcelable
     }
 
     public void setHttpHeaders(JSONObject httpHeaders) throws JSONException {
+        // intentionally set httpHeaders to empty hash map
+        // this allows to reset headers in .fromJSONArray providing empty httpHeaders JSONObject
         this.httpHeaders = new HashMap<String, String>();
         if (httpHeaders == null) {
             return;
@@ -282,28 +317,39 @@ public class Config implements Parcelable
         }
     }
 
+    public Integer getMaxLocations() {
+        return maxLocations;
+    }
+
+    public void setMaxLocations(Integer maxLocations) {
+        this.maxLocations = maxLocations;
+    }
+
     @Override
     public String toString () {
         return new StringBuffer()
                 .append("Config[distanceFilter=").append(getDistanceFilter())
                 .append(" stationaryRadius=").append(getStationaryRadius())
                 .append(" desiredAccuracy=").append(getDesiredAccuracy())
-                .append(" isDebugging=").append(isDebugging())
-                .append(" stopOnTerminate=" ).append(getStopOnTerminate())
-                .append(" url=").append(getUrl())
-                .append(" httpHeaders=").append(getHttpHeaders().toString())
                 .append(" interval=").append(getInterval())
                 .append(" fastestInterval=").append(getFastestInterval())
                 .append(" activitiesInterval=").append(getActivitiesInterval())
+                .append(" isDebugging=").append(isDebugging())
+                .append(" stopOnTerminate=" ).append(getStopOnTerminate())
+                .append(" stopOnStillActivity=").append(getStopOnStillActivity())
+                .append(" startOnBoot=").append(getStartOnBoot())
+                .append(" startForeground=").append(getStartForeground())
+                .append(" locationProvider=").append(getLocationProvider())
                 .append(" nTitle=").append(getNotificationTitle())
                 .append(" nText=").append(getNotificationText())
                 .append(" nIconLarge=").append(getLargeNotificationIcon())
                 .append(" nIconSmall=").append(getSmallNotificationIcon())
                 .append(" nIconColor=").append(getNotificationIconColor())
-                .append(" startOnBoot=").append(getStartOnBoot())
-                .append(" startForeground=").append(getStartForeground())
-                .append(" locationProvider=").append(getLocationProvider())
-                .append(" stopOnStillActivity=").append(getStopOnStillActivity())
+                .append(" url=").append(getUrl())
+                .append(" syncUrl=").append(getSyncUrl())
+                .append(" syncThreshold=").append(getSyncThreshold())
+                .append(" httpHeaders=").append(getHttpHeaders().toString())
+                .append(" maxLocations=").append(getMaxLocations())
                 .append("]")
                 .toString();
     }
@@ -343,7 +389,10 @@ public class Config implements Parcelable
         config.setStartForeground(jObject.optBoolean("startForeground", config.getStartForeground()));
         config.setStopOnStillActivity(jObject.optBoolean("stopOnStillActivity", config.getStopOnStillActivity()));
         config.setUrl(jObject.optString("url"));
+        config.setSyncUrl(jObject.optString("syncUrl"));
+        config.setSyncThreshold(jObject.optInt("syncThreshold", config.getSyncThreshold()));
         config.setHttpHeaders(jObject.optJSONObject("httpHeaders"));
+        config.setMaxLocations(jObject.optInt("maxLocations", config.getMaxLocations()));
 
         return config;
     }
@@ -368,7 +417,10 @@ public class Config implements Parcelable
         json.put("activitiesInterval", getActivitiesInterval());
         json.put("stopOnStillActivity", getStopOnStillActivity());
         json.put("url", getUrl());
+        json.put("syncUrl", getSyncUrl());
+        json.put("syncThreshold", getSyncThreshold());
         json.put("httpHeaders", new JSONObject(getHttpHeaders()));
+        json.put("maxLocations", getMaxLocations());
 
         return json;
   	}