+++ android/plugin/src/main/java/com/marianhello/bgloc/data/LocationDAO.java
@@ -4,7 +4,9 @@ import java.util.Collection;
 
 public interface LocationDAO {
     public Collection<BackgroundLocation> getAllLocations();
-    public Long persistLocation(BackgroundLocation l);
+    public Collection<BackgroundLocation> getValidLocations();
+    public Long persistLocation(BackgroundLocation location);
+    public Long persistLocationWithLimit(BackgroundLocation location, Integer maxRows);
     public void deleteLocation(Long locationId);
     public void deleteAllLocations();
 }