+++ android/plugin/src/main/java/com/marianhello/bgloc/data/sqlite/LocationContract.java
@@ -21,5 +21,6 @@ public final class LocationContract {
         public static final String COLUMN_NAME_PROVIDER = "provider";
         public static final String COLUMN_NAME_LOCATION_PROVIDER = "service_provider";
         public static final String COLUMN_NAME_DEBUG = "debug";
+        public static final String COLUMN_NAME_VALID = "valid";
     }
 }