+++ android/plugin/src/main/java/com/marianhello/bgloc/data/sqlite/SQLiteConfigurationDAO.java
@@ -23,16 +23,20 @@ import com.marianhello.bgloc.data.sqlite.ConfigurationContract.ConfigurationEntr
 import com.tenforwardconsulting.cordova.BackgroundGeolocationPlugin;
 
 public class SQLiteConfigurationDAO implements ConfigurationDAO {
-  public static final String DATE_FORMAT = "yyyy-MM-dd'T'HH:mm'Z'";
   private static final String TAG = SQLiteConfigurationDAO.class.getSimpleName();
-  private Context context;
+
+    private SQLiteDatabase db;
 
   public SQLiteConfigurationDAO(Context context) {
-      this.context = context;
+    SQLiteOpenHelper helper = SQLiteOpenHelper.getHelper(context);
+    this.db = helper.getWritableDatabase();
+  }
+
+  public SQLiteConfigurationDAO(SQLiteDatabase db) {
+    this.db = db;
   }
 
   public Config retrieveConfiguration() throws JSONException {
-    SQLiteDatabase db = null;
     Cursor cursor = null;
 
     String[] columns = {
@@ -47,6 +51,7 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
       ConfigurationEntry.COLUMN_NAME_NOTIF_ICON_SMALL,
       ConfigurationEntry.COLUMN_NAME_NOTIF_COLOR,
       ConfigurationEntry.COLUMN_NAME_STOP_TERMINATE,
+      ConfigurationEntry.COLUMN_NAME_STOP_ON_STILL,
       ConfigurationEntry.COLUMN_NAME_START_BOOT,
       ConfigurationEntry.COLUMN_NAME_START_FOREGROUND,
       ConfigurationEntry.COLUMN_NAME_LOCATION_PROVIDER,
@@ -54,7 +59,10 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
       ConfigurationEntry.COLUMN_NAME_FASTEST_INTERVAL,
       ConfigurationEntry.COLUMN_NAME_ACTIVITIES_INTERVAL,
       ConfigurationEntry.COLUMN_NAME_URL,
-      ConfigurationEntry.COLUMN_NAME_HEADERS
+      ConfigurationEntry.COLUMN_NAME_SYNC_URL,
+      ConfigurationEntry.COLUMN_NAME_SYNC_THRESHOLD,
+      ConfigurationEntry.COLUMN_NAME_HEADERS,
+      ConfigurationEntry.COLUMN_NAME_MAX_LOCATIONS
     };
 
     String whereClause = null;
@@ -65,7 +73,6 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
 
     Config config = null;
     try {
-      db = new SQLiteOpenHelper(context).getReadableDatabase();
       cursor = db.query(
           ConfigurationEntry.TABLE_NAME,  // The table to query
           columns,                   // The columns to return
@@ -82,22 +89,17 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
       if (cursor != null) {
         cursor.close();
       }
-      if (db != null) {
-        db.close();
-      }
     }
     return config;
   }
 
   public boolean persistConfiguration(Config config) throws NullPointerException {
-    SQLiteDatabase db = new SQLiteOpenHelper(context).getWritableDatabase();
     db.beginTransaction();
     db.delete(ConfigurationEntry.TABLE_NAME, null, null);
     long rowId = db.insert(ConfigurationEntry.TABLE_NAME, ConfigurationEntry.COLUMN_NAME_NULLABLE, getContentValues(config));
-    Log.d(TAG, "After insert, rowId = " + rowId);
+    Log.d(TAG, "Configuration persisted with rowId = " + rowId);
     db.setTransactionSuccessful();
     db.endTransaction();
-    db.close();
     if (rowId > -1) {
       return true;
     } else {
@@ -117,6 +119,7 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
     config.setLargeNotificationIcon(c.getString(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_NOTIF_ICON_LARGE)));
     config.setNotificationIconColor(c.getString(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_NOTIF_COLOR)));
     config.setStopOnTerminate( (c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_STOP_TERMINATE)) == 1) ? true : false );
+    config.setStopOnStillActivity( (c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_STOP_ON_STILL)) == 1) ? true : false );
     config.setStartOnBoot( (c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_START_BOOT)) == 1) ? true : false );
     config.setStartForeground( (c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_START_FOREGROUND)) == 1) ? true : false );
     config.setLocationProvider(c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_LOCATION_PROVIDER)));
@@ -124,7 +127,10 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
     config.setFastestInterval(c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_FASTEST_INTERVAL)));
     config.setActivitiesInterval(c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_ACTIVITIES_INTERVAL)));
     config.setUrl(c.getString(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_URL)));
+    config.setSyncUrl(c.getString(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_SYNC_URL)));
+    config.setSyncThreshold(c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_SYNC_THRESHOLD)));
     config.setHttpHeaders(new JSONObject(c.getString(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_HEADERS))));
+    config.setMaxLocations(c.getInt(c.getColumnIndex(ConfigurationEntry.COLUMN_NAME_MAX_LOCATIONS)));
 
     return config;
   }
@@ -141,6 +147,7 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
     values.put(ConfigurationEntry.COLUMN_NAME_NOTIF_ICON_LARGE, config.getLargeNotificationIcon());
     values.put(ConfigurationEntry.COLUMN_NAME_NOTIF_COLOR, config.getNotificationIconColor());
     values.put(ConfigurationEntry.COLUMN_NAME_STOP_TERMINATE, (config.getStopOnTerminate() == true) ? 1 : 0);
+    values.put(ConfigurationEntry.COLUMN_NAME_STOP_ON_STILL, (config.getStopOnStillActivity() == true) ? 1 : 0);
     values.put(ConfigurationEntry.COLUMN_NAME_START_BOOT, (config.getStartOnBoot() == true) ? 1 : 0);
     values.put(ConfigurationEntry.COLUMN_NAME_START_FOREGROUND, (config.getStartForeground() == true) ? 1 : 0);
     values.put(ConfigurationEntry.COLUMN_NAME_LOCATION_PROVIDER, config.getLocationProvider());
@@ -148,7 +155,10 @@ public class SQLiteConfigurationDAO implements ConfigurationDAO {
     values.put(ConfigurationEntry.COLUMN_NAME_FASTEST_INTERVAL, config.getFastestInterval());
     values.put(ConfigurationEntry.COLUMN_NAME_ACTIVITIES_INTERVAL, config.getActivitiesInterval());
     values.put(ConfigurationEntry.COLUMN_NAME_URL, config.getUrl());
+    values.put(ConfigurationEntry.COLUMN_NAME_SYNC_URL, config.getSyncUrl());
+    values.put(ConfigurationEntry.COLUMN_NAME_SYNC_THRESHOLD, config.getSyncThreshold());
     values.put(ConfigurationEntry.COLUMN_NAME_HEADERS, new JSONObject(config.getHttpHeaders()).toString());
+    values.put(ConfigurationEntry.COLUMN_NAME_MAX_LOCATIONS, config.getMaxLocations());
 
     return values;
   }