+++ android/plugin/src/main/java/com/marianhello/bgloc/data/sqlite/SQLiteOpenHelper.java
@@ -2,15 +2,21 @@ package com.marianhello.bgloc.data.sqlite;
 
 
 import android.content.Context;
+import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
+import android.text.TextUtils;
 import android.util.Log;
 
 import com.marianhello.bgloc.data.sqlite.LocationContract.LocationEntry;
 import com.marianhello.bgloc.data.sqlite.ConfigurationContract.ConfigurationEntry;
 
+import java.util.Arrays;
+import java.util.List;
+import java.util.StringTokenizer;
+
 public class SQLiteOpenHelper extends android.database.sqlite.SQLiteOpenHelper {
-    private static final String SQLITE_DATABASE_NAME = "cordova_bg_geolocation.db";
-    private static final int DATABASE_VERSION = 10;
+    public static final String SQLITE_DATABASE_NAME = "cordova_bg_geolocation.db";
+    public static final int DATABASE_VERSION = 11;
     private static final String TEXT_TYPE = " TEXT";
     private static final String INTEGER_TYPE = " INTEGER";
     private static final String REAL_TYPE = " REAL";
@@ -28,7 +34,8 @@ public class SQLiteOpenHelper extends android.database.sqlite.SQLiteOpenHelper {
         LocationEntry.COLUMN_NAME_LONGITUDE + REAL_TYPE + COMMA_SEP +
         LocationEntry.COLUMN_NAME_PROVIDER + TEXT_TYPE + COMMA_SEP +
         LocationEntry.COLUMN_NAME_LOCATION_PROVIDER + INTEGER_TYPE + COMMA_SEP +
-        LocationEntry.COLUMN_NAME_DEBUG + INTEGER_TYPE +
+        LocationEntry.COLUMN_NAME_DEBUG + INTEGER_TYPE + COMMA_SEP +
+        LocationEntry.COLUMN_NAME_VALID + INTEGER_TYPE +
         " )";
 
     private static final String SQL_CREATE_CONFIG_TABLE =
@@ -44,6 +51,7 @@ public class SQLiteOpenHelper extends android.database.sqlite.SQLiteOpenHelper {
         ConfigurationEntry.COLUMN_NAME_NOTIF_ICON_LARGE + TEXT_TYPE + COMMA_SEP +
         ConfigurationEntry.COLUMN_NAME_NOTIF_COLOR + TEXT_TYPE + COMMA_SEP +
         ConfigurationEntry.COLUMN_NAME_STOP_TERMINATE + INTEGER_TYPE + COMMA_SEP +
+        ConfigurationEntry.COLUMN_NAME_STOP_ON_STILL + INTEGER_TYPE + COMMA_SEP +
         ConfigurationEntry.COLUMN_NAME_START_BOOT + INTEGER_TYPE + COMMA_SEP +
         ConfigurationEntry.COLUMN_NAME_START_FOREGROUND + INTEGER_TYPE + COMMA_SEP +
         ConfigurationEntry.COLUMN_NAME_LOCATION_PROVIDER + TEXT_TYPE + COMMA_SEP +
@@ -51,40 +59,106 @@ public class SQLiteOpenHelper extends android.database.sqlite.SQLiteOpenHelper {
         ConfigurationEntry.COLUMN_NAME_FASTEST_INTERVAL + INTEGER_TYPE + COMMA_SEP +
         ConfigurationEntry.COLUMN_NAME_ACTIVITIES_INTERVAL + INTEGER_TYPE + COMMA_SEP +
         ConfigurationEntry.COLUMN_NAME_URL + TEXT_TYPE + COMMA_SEP +
-        ConfigurationEntry.COLUMN_NAME_HEADERS + TEXT_TYPE +
+        ConfigurationEntry.COLUMN_NAME_SYNC_URL + TEXT_TYPE + COMMA_SEP +
+        ConfigurationEntry.COLUMN_NAME_SYNC_THRESHOLD + INTEGER_TYPE + COMMA_SEP +
+        ConfigurationEntry.COLUMN_NAME_HEADERS + TEXT_TYPE + COMMA_SEP +
+        ConfigurationEntry.COLUMN_NAME_MAX_LOCATIONS + INTEGER_TYPE +
         " )";
 
-    private static final String SQL_DELETE_CONFIG_TABLE =
+    private static final String SQL_DROP_CONFIG_TABLE =
             "DROP TABLE IF EXISTS " + ConfigurationEntry.TABLE_NAME;
 
-    private static final String SQL_DELETE_LOCATION_TABLE =
+    private static final String SQL_DROP_LOCATION_TABLE =
             "DROP TABLE IF EXISTS " + LocationEntry.TABLE_NAME;
 
-    SQLiteOpenHelper(Context context) {
+    private static final String SQL_CREATE_LOCATION_TABLE_TIME_IDX =
+            "CREATE INDEX time_idx ON " + LocationEntry.TABLE_NAME + " (" + LocationEntry.COLUMN_NAME_TIME + ")";
+
+    private static SQLiteOpenHelper instance;
+
+    /**
+     * Get SqliteOpenHelper instance (singleton)
+     *
+     * Use the application context, which will ensure that you
+     * don't accidentally leak an Activity's context.
+     * See this article for more information: http://bit.ly/6LRzfx
+     *
+     * @param context
+     * @return
+     */
+    public static synchronized SQLiteOpenHelper getHelper(Context context) {
+        if (instance == null)
+            instance = new SQLiteOpenHelper(context);
+
+        return instance;
+    }
+
+    /**
+     * Constructor
+     *
+     * NOTE: Intended to use only for testing purposes.
+     * Use factory method getHelper instead.
+     *
+     * @param context
+     */
+    public SQLiteOpenHelper(Context context) {
         super(context, SQLITE_DATABASE_NAME, null, DATABASE_VERSION);
     }
 
     @Override
     public void onCreate(SQLiteDatabase db) {
+        Log.d(this.getClass().getName(), "Creating db: " + this.getDatabaseName());
         db.execSQL(SQL_CREATE_LOCATION_TABLE);
         Log.d(this.getClass().getName(), SQL_CREATE_LOCATION_TABLE);
         db.execSQL(SQL_CREATE_CONFIG_TABLE);
         Log.d(this.getClass().getName(), SQL_CREATE_CONFIG_TABLE);
+        db.execSQL(SQL_CREATE_LOCATION_TABLE_TIME_IDX);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        // This database is only a cache for online data, so its upgrade policy is
-        // to simply to discard the data and start over
-        db.execSQL(SQL_DELETE_LOCATION_TABLE);
-        Log.d(this.getClass().getName(), SQL_DELETE_LOCATION_TABLE);
-        db.execSQL(SQL_DELETE_CONFIG_TABLE);
-        Log.d(this.getClass().getName(), SQL_DELETE_CONFIG_TABLE);
-        onCreate(db);
+        if (oldVersion == 10) {
+            Log.d(this.getClass().getName(), "Upgrading database oldVersion: " + oldVersion + " newVersion: " + newVersion);
+            String alterSql[] = {
+                    "ALTER TABLE " + LocationEntry.TABLE_NAME +
+                            " ADD COLUMN " + LocationEntry.COLUMN_NAME_VALID + INTEGER_TYPE,
+//                    "ALTER TABLE " + ConfigurationEntry.TABLE_NAME +
+//                            " ADD COLUMN " + ConfigurationEntry.COLUMN_NAME_STOP_ON_STILL + INTEGER_TYPE,
+//                    "ALTER TABLE " + ConfigurationEntry.TABLE_NAME +
+//                            " ADD COLUMN " + ConfigurationEntry.COLUMN_NAME_MAX_LOCATIONS + INTEGER_TYPE,
+//                    "ALTER TABLE " + ConfigurationEntry.TABLE_NAME +
+//                            " ADD COLUMN " + ConfigurationEntry.COLUMN_NAME_SYNC_URL + TEXT_TYPE,
+//                    "ALTER TABLE " + ConfigurationEntry.TABLE_NAME +
+//                            " ADD COLUMN " + ConfigurationEntry.COLUMN_NAME_SYNC_THRESHOLD + INTEGER_TYPE,
+//                    "UPDATE " + ConfigurationEntry.TABLE_NAME +
+//                            " SET " + ConfigurationEntry.COLUMN_NAME_MAX_LOCATIONS + "= 10000",
+//                    "UPDATE " + ConfigurationEntry.TABLE_NAME +
+//                            " SET " + ConfigurationEntry.COLUMN_NAME_SYNC_THRESHOLD + "= 100",
+                    SQL_CREATE_LOCATION_TABLE_TIME_IDX,
+                    SQL_DROP_CONFIG_TABLE,
+                    SQL_CREATE_CONFIG_TABLE
+            };
+
+            for (String sql : alterSql) {
+                db.execSQL(sql);
+            }
+        } else {
+            // for all other scenarios drop table and start over
+            db.execSQL(SQL_DROP_LOCATION_TABLE);
+            Log.d(this.getClass().getName(), SQL_DROP_LOCATION_TABLE);
+            db.execSQL(SQL_DROP_CONFIG_TABLE);
+            Log.d(this.getClass().getName(), SQL_DROP_CONFIG_TABLE);
+            onCreate(db);
+        }
     }
 
     @Override
     public void onDowngrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        onUpgrade(db, oldVersion, newVersion);
+        // we don't support db downgrade yet, instead we drop table and start over
+        db.execSQL(SQL_DROP_LOCATION_TABLE);
+        Log.d(this.getClass().getName(), SQL_DROP_LOCATION_TABLE);
+        db.execSQL(SQL_DROP_CONFIG_TABLE);
+        Log.d(this.getClass().getName(), SQL_DROP_CONFIG_TABLE);
+        onCreate(db);
     }
 }